Adjustments in project files as a result of renaming repository and projects to get rid of old project names
