+++ .vscode/launch.json
@@ -10,9 +10,9 @@
             "request": "launch",
             "preLaunchTask": "build",
             // If you have changed target frameworks, make sure to update the program path.
-            "program": "${workspaceFolder}/PalindromicLibTests/bin/Debug/net6.0/PalindromicLibTests.dll",
+            "program": "${workspaceFolder}/PalindromesTests/bin/Debug/net6.0/PalindromesTests.dll",
             "args": [],
-            "cwd": "${workspaceFolder}/PalindromicLibTests",
+            "cwd": "${workspaceFolder}/PalindromesTests",
             // For more information about the 'console' field, see https://aka.ms/VSCode-CS-LaunchJson-Console
             "console": "internalConsole",
             "stopAtEntry": false
