+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "configuration": "release",
     "ConsoleApp.branchCoverage.threshold": 90,
-    "PalindromicLib.branchCoverage.threshold": 90,
+    "Palindromes.branchCoverage.threshold": 90,
     "debug.terminal.clearBeforeReusing": true,
     "terminal.integrated.automationProfile.windows": {
         "path": "C:\\Windows\\System32\\cmd.exe"
