+++ ConsoleApp/ConsoleApp.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\PalindromicLib\Palindromes.csproj" />
+    <ProjectReference Include="..\Palindromes\Palindromes.csproj" />
   </ItemGroup>
 
 </Project>
