+++ ConsoleApp/FileParserForChallenge12050.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 
-using PalindromicLib;
+using Palindromes;
 
 namespace ConsoleApp;
