+++ Palindromes.sln
@@ -1,0 +1,43 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34031.279
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Palindromes", "Palindromes\Palindromes.csproj", "{00DFCCE8-469D-4BFA-BF58-821B24BDA73C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PalindromesTests", "PalindromesTests\PalindromesTests.csproj", "{5B70131B-10D8-4A46-923D-73B64DA80F17}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp", "ConsoleApp\ConsoleApp.csproj", "{8A8B532D-518D-4FFA-910D-BE659CA5A086}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleAppTests", "ConsoleAppTests\ConsoleAppTests.csproj", "{463CBB6B-84DE-472A-A44E-783C3A36206F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{00DFCCE8-469D-4BFA-BF58-821B24BDA73C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{00DFCCE8-469D-4BFA-BF58-821B24BDA73C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{00DFCCE8-469D-4BFA-BF58-821B24BDA73C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{00DFCCE8-469D-4BFA-BF58-821B24BDA73C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B70131B-10D8-4A46-923D-73B64DA80F17}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5B70131B-10D8-4A46-923D-73B64DA80F17}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5B70131B-10D8-4A46-923D-73B64DA80F17}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5B70131B-10D8-4A46-923D-73B64DA80F17}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8A8B532D-518D-4FFA-910D-BE659CA5A086}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8A8B532D-518D-4FFA-910D-BE659CA5A086}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8A8B532D-518D-4FFA-910D-BE659CA5A086}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8A8B532D-518D-4FFA-910D-BE659CA5A086}.Release|Any CPU.Build.0 = Release|Any CPU
+		{463CBB6B-84DE-472A-A44E-783C3A36206F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{463CBB6B-84DE-472A-A44E-783C3A36206F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{463CBB6B-84DE-472A-A44E-783C3A36206F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{463CBB6B-84DE-472A-A44E-783C3A36206F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9D1E82EF-6BD2-4A1C-B931-C51798C54302}
+	EndGlobalSection
+EndGlobal
