+++ Palindromes/Palindromes.csproj
