+++ PalindromesTests/StrykerOutput/2023-09-19.18-34-50/logs/log-20230919.txt
