+++ PalindromesTests/StrykerOutput/2023-09-19.22-15-56/logs/log-20230919.txt
