+++ PalindromesTests/StrykerOutput/2023-09-20.11-12-28/logs/log-20230920.txt
