+++ PalindromesTests/StrykerOutput/2023-09-20.12-24-24/logs/log-20230920.txt
