+++ PalindromesTests/StrykerOutput/2023-09-23.14-14-44/logs/log-20230923.txt
