+++ PalindromesTests/stryker-partial-config.json
@@ -2,21 +2,22 @@
     "stryker-config": {
         "additional-timeout": 30000,
         "baseline": {},
+        "since": {
+            "enabled": true,
+            "target": "main"
+        },
         "concurrency": 2,
-        "since": {},
         "coverage-analysis": "perTest",
         "disable-bail": false,
         "disable-mix-mutants": false,
         "ignore-methods": [],
         "ignore-mutations": [],
         "language-version": "Csharp10",
-        "mutate": [
-            "!**/Program.cs"
-        ],
+        "mutate": [],
         "mutation-level": "Complete",
-        "project": "ConsoleApp\\ConsoleApp.csproj",
+        "project": "Palindromes\\Palindromes.csproj",
         "project-info": {},
-        "report-file-name": "stryker-full-report.html",
+        "report-file-name": "stryker-partial-report.html",
         "reporters": [
             "progress",
             "html",
@@ -26,7 +27,7 @@
         "target-framework": "net6.0",
         "test-case-filter": "",
         "test-projects": [
-            "ConsoleAppTests.csproj"
+            "PalindromesTests.csproj"
         ],
         "thresholds": {
             "high": 80,
